Revert "move createTestBankAccount to user SDK"
This reverts commit 26bc9e8e551516d49472a28b6b0ed1e56b6cd091.
