+++ common/src/main/java/io/token/Member.java
@@ -806,6 +806,50 @@ public class Member {
         client.clearTrackingMetadata();
     }
 
+    /**
+     * Creates a test bank account in a fake bank.
+     *
+     * @param balance account balance to set
+     * @param currency currency code, e.g. "EUR"
+     * @return OAuth bank authorization
+     */
+    public Observable<OauthBankAuthorization> createTestBankAccount(
+            double balance, String currency) {
+        return client.createTestBankAccount(MoneyProtos.Money.newBuilder()
+                .setCurrency(currency)
+                .setValue(Double.toString(balance))
+                .build());
+    }
+
+    /**
+     * Creates a test bank account in a fake bank and links the account.
+     *
+     * @param balance account balance to set
+     * @param currency currency code, e.g. "EUR"
+     * @return the linked account
+     */
+    protected Observable<Account> createAndLinkTestBankAccountImpl(
+            double balance,
+            String currency) {
+        return toAccount(client.createAndLinkTestBankAccount(
+                MoneyProtos.Money.newBuilder()
+                        .setValue(Double.toString(balance))
+                        .setCurrency(currency)
+                        .build()
+        ));
+    }
+
+    /**
+     * Creates a test bank account in a fake bank.
+     *
+     * @param balance account balance to set
+     * @param currency currency code, e.g. "EUR"
+     * @return OAuth bank authorization
+     */
+    public OauthBankAuthorization createTestBankAccountBlocking(double balance, String currency) {
+        return createTestBankAccount(balance, currency).blockingSingle();
+    }
+
     @Override
     public int hashCode() {
         return client.hashCode();
